Owner API all done

+++ Bomholt.PetShop.RestAPI/Controllers/OwnersController.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using Bomholt.PetShop.Core.ApplicationService;
 using Bomholt.PetShop.Core.Entities;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Bomholt.PetShop.RestAPI.Controllers
@@ -29,31 +25,72 @@ namespace Bomholt.PetShop.RestAPI.Controllers
 
         // GET: api/Owners/5
         [HttpGet("{id}", Name = "Get")]
-        public Owner Get(int id)
+        public ActionResult Get(int id)
         {
-            return _ownerService.GetById(id);
+            if (id < 1)
+            {
+                return BadRequest("Id must be lager than zero!");
+            }
+            Owner OwnerFound = _ownerService.GetById(id);
+            if (OwnerFound == null)
+            {
+                return NotFound($"No owner with id {id} found!");
+            }
+            return Ok(OwnerFound);
         }
         
         // POST: api/Owners
         [HttpPost]
-        public void Post([FromBody]Owner value)
+        public ActionResult Post([FromBody]Owner value)
         {
-            _ownerService.CreateNew(value);
+            bool success = _ownerService.CreateNew(value);
+            if (success)
+            {
+                return Ok($"Owner {value} was successfully created");
+            }
+            else
+            {
+                return NotFound("Something went wrong!");
+            }
         }
         
         // PUT: api/Owners/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody]Owner value)
+        public ActionResult Put(int id, [FromBody]Owner value)
         {
             value.ID = id;
-            _ownerService.Update(value);
+            if (id < 1)
+            {
+                return BadRequest("Id must be lager than zero!");
+            }
+            bool success = _ownerService.Update(value);
+            if (success)
+            {
+                return Ok($"Owner nr. {id} was updated");
+            }
+            else
+            {
+                return NotFound($"No owner found with id {id}");
+            }
         }
         
         // DELETE: api/ApiWithActions/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public ActionResult Delete(int id)
         {
-            _ownerService.DeleteById(id);
+            if (id < 1)
+            {
+                return BadRequest("Id must be lager than zero!");
+            }
+            bool success = _ownerService.DeleteById(id);
+            if (success)
+            {
+                return Ok("Owner nr. {id} was deleted");
+            }
+            else
+            {
+                return NotFound("No owner found with id {id}");
+            }
         }
     }
 }
