+++ Bomholt.PetShop.RestAPI/Controllers/PetsController.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using Bomholt.PetShop.Core.ApplicationService;
 using Bomholt.PetShop.Core.Entities;
 using Microsoft.AspNetCore.Mvc;
@@ -27,31 +24,72 @@ namespace Bomholt.PetShop.RestAPI.Controllers
 
         // GET api/pets/5
         [HttpGet("{id}")]
-        public Pet Get(int id)
+        public ActionResult Get(int id)
         {
-            return _petService.GetById(id);
+            if (id < 1)
+            {
+                return BadRequest("Id must be lager than zero!");
+            }
+            Pet PetFound = _petService.GetById(id);
+            if (PetFound == null)
+            {
+                return NotFound($"No pet with id {id} found!");
+            }
+            return Ok(PetFound);
         }
 
         // POST api/pets
         [HttpPost]
-        public void Post([FromBody]Pet value)
+        public ActionResult Post([FromBody]Pet value)
         {
             bool success = _petService.CreateNewPet(value);
+            if (success)
+            {
+                return Ok($"Pet {value} was created");
+            }
+            else
+            {
+                return NotFound("Something went wrong!");
+            }
         }
 
         // PUT api/pets/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody]Pet value)
+        public ActionResult Put(int id, [FromBody]Pet value)
         {
             value.ID = id;
+            if (id < 1)
+            {
+                return BadRequest("Id must be lager than zero!");
+            }
             bool success = _petService.UpdatePet(value);
+            if (success)
+            {
+                return Ok($"Pet nr. {id} was updated");
+            }
+            else
+            {
+                return NotFound($"No pet found with id {id}");
+            }
         }
 
         // DELETE api/pets/5
         [HttpDelete("{id}")]
-        public bool Delete(int id)
+        public ActionResult Delete(int id)
         {
-            return _petService.DeletePetById(id);
+            if (id < 1)
+            {
+                return BadRequest("Id must be lager than zero!");
+            }
+            bool success = _petService.DeletePetById(id);
+            if (success)
+            {
+                return Ok($"Pet nr. {id} was deleted");
+            }
+            else
+            {
+                return NotFound($"No pet found with id {id}");
+            }
         }
     }
 }
